Bug fixes and simplify comments
